+++ entrypoint.sh
@@ -20,9 +20,6 @@ if [ "$GPU" != "cpu" ]; then
             cuda)
                 pip install --target "$CACHE_DIR" torch==2.11.0 torchaudio==2.11.0 torchvision==0.26.0 onnxruntime-gpu==1.26.0
                 ;;
-            rocm)
-                pip install --target "$CACHE_DIR" torch==2.11.0+rocm7.1 torchaudio==2.11.0+rocm7.1 torchvision==0.26.0+rocm7.1 onnxruntime --extra-index-url https://download.pytorch.org/whl/rocm7.1
-                ;;
         esac
         echo "✅ $GPU backend installed"
     fi
@@ -34,9 +31,6 @@ if [ "$GPU" != "cpu" ]; then
         case $GPU in
             cuda)
                 pip install --target "$CACHE_DIR" onnxruntime-gpu==1.26.0
-                ;;
-            rocm)
-                pip install --target "$CACHE_DIR" onnxruntime
                 ;;
         esac
         echo "✅ onnxruntime reinstalled"